refactor: reorganize prompt components and add dynamic few-shot prompt functionality
