--- ragas/src/ragas/experimental/prompt/__init__.py
@@ -1,4 +1,0 @@
-from .base import Prompt
-from .dynamic_few_shot import DynamicFewShotPrompt
-
-__all__ = ["Prompt", "DynamicFewShotPrompt"]
