+++ ragas/src/ragas/prompt/__init__.py
@@ -7,6 +7,12 @@ from .few_shot_pydantic_prompt import (
 from .mixin import PromptMixin
 from .multi_modal_prompt import ImageTextPrompt, ImageTextPromptValue
 from .pydantic_prompt import InputModel, OutputModel, PydanticPrompt
+from .simple_prompt import Prompt
+from .dynamic_few_shot import (
+    DynamicFewShotPrompt,
+    SimpleExampleStore,
+    SimpleInMemoryExampleStore,
+)
 
 __all__ = [
     "BasePrompt",
@@ -22,4 +28,8 @@ __all__ = [
     "OutputModel",
     "ImageTextPrompt",
     "ImageTextPromptValue",
+    "Prompt",
+    "DynamicFewShotPrompt",
+    "SimpleExampleStore",
+    "SimpleInMemoryExampleStore",
 ]
