+++ ragas/tests/experimental/unit/test_prompt_save_load.py
@@ -4,7 +4,7 @@ import warnings
 import pytest
 from pydantic import BaseModel
 
-from ragas.experimental.prompt.base import Prompt
+from ragas.prompt import Prompt
 
 
 class MockResponseModel(BaseModel):
